Move i18n app to settings (discoverability)
