+++ I18N.md
@@ -7,7 +7,7 @@ The apps UI allows all strings to be translated. Additionally it has a basic UI 
 To update translations, the following process is required.
 
 - launch the apps UI, either locally or via [https://polkadot.js.org/apps](https://polkadot.js.org/apps)
-- explicitly navigate to the i18n page, [https://polkadot.js.org/apps/#/i18n](https://polkadot.js.org/apps/#/i18n)
+- explicitly navigate to the i18n page, [https://polkadot.js.org/apps/#/i18n](https://polkadot.js.org/apps/#/settings/i18n)
 
 Here you will find a dropdown of all the available languages and all the modules that maps to the UI. On a single screen you will be able to see all the available strings for a specific module.
 
@@ -17,4 +17,6 @@ Here you will find a dropdown of all the available languages and all the modules
 
 ## Adding a new language
 
-The process is similar for the above, but does require a new folder with the language identifier to be added. Create `packages/apps/public/locales/<id>` folder with an empty `translation.json` (containing only `{}`). After addition of the folder, run `yarn build:i18n` and then the new language will be available for update as per the process in the previous section.
+The process is similar for the above, but does require a new folder with the language identifier to be added. Create [packages/apps/public/locales/<id>](https://github.com/polkadot-js/apps/tree/master/packages/apps/public/locales) folder with an empty `translation.json` (containing only `{}`). After addition of the folder, run `yarn build:i18n` and then the new language will be available for update as per the process in the previous section.
+
+In addition to the language folder, the language also needs to be added to the dropdown for available languages, this can be found in [packages/apps-config/src/settings.languages.ts](https://github.com/polkadot-js/apps/blob/master/packages/apps-config/src/settings/languages.ts)
