+++ packages/apps/public/locales/en/translation.json
@@ -566,6 +566,7 @@
   "To council": "",
   "To ensure optimal fund security using the same stash/controller is strongly discouraged, but not forbidden.": "",
   "Transfer": "",
+  "Translate": "",
   "Treasury overview": "",
   "Type here what you would like your address to contain. This tool will generate the keys and show the associated addresses that best match your search. You can use \"?\" as a wildcard for a character.": "",
   "Type the amount you want to transfer. Note that you can select the unit on the right e.g sending 1 milli is equivalent to sending 0.001.": "",
