--- packages/page-i18n/src/translate.ts
@@ -1,9 +1,0 @@
-// Copyright 2017-2020 @polkadot/app-i18n authors & contributors
-// This software may be modified and distributed under the terms
-// of the Apache-2.0 license. See the LICENSE file for details.
-
-import { useTranslation as useTranslationBase, UseTranslationResponse } from 'react-i18next';
-
-export function useTranslation (): UseTranslationResponse {
-  return useTranslationBase('app-i18n');
-}
