+++ packages/page-settings/src/I18n/StringInput.tsx
@@ -1,4 +1,4 @@
-// Copyright 2017-2020 @polkadot/app-i18n authors & contributors
+// Copyright 2017-2020 @polkadot/app-settings authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
