+++ packages/page-settings/src/index.tsx
@@ -12,6 +12,7 @@ import uiSettings from '@polkadot/ui-settings';
 import md from './md/basics.md';
 import { useTranslation } from './translate';
 import Developer from './Developer';
+import I18n from './I18n';
 import Metadata from './Metadata';
 import General from './General';
 import useCounter from './useCounter';
@@ -42,6 +43,10 @@ function SettingsApp ({ basePath, onStatusChange }: Props): React.ReactElement<P
     {
       name: 'developer',
       text: t('Developer')
+    },
+    {
+      name: 'i18n',
+      text: t('Translate')
     }
   ], [numExtensions, t]);
 
@@ -61,6 +66,9 @@ function SettingsApp ({ basePath, onStatusChange }: Props): React.ReactElement<P
             basePath={basePath}
             onStatusChange={onStatusChange}
           />
+        </Route>
+        <Route path={`${basePath}/i18n`}>
+          <I18n />
         </Route>
         <Route path={`${basePath}/metadata`}>
           <Metadata />
