+++ packages/react-components/src/i18n/cache.ts
@@ -1,0 +1,7 @@
+// Copyright 2017-2020 @polkadot/react-components authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+const languageCache: Record<string, Record<string, string>> = {};
+
+export default languageCache;
